# Implement custom easing functions