… osu.Framework/Graphics/TransformSequenceExtensions.cs
… osu.Framework/Graphics/TransformableExtensions.cs
… osu.Framework/Graphics/Transforms/IEasingFunction.cs
… osu.Framework/Graphics/Transforms/Transform.cs
… osu.Framework/Graphics/Transforms/TransformBindable.cs
… osu.Framework/Graphics/Transforms/TransformCustom.cs
… osu.Framework/Graphics/Transforms/Transformable.cs
… osu.Framework/Graphics/UserInterface/CircularProgress.cs
… osu.Framework/Utils/IInterpolable.cs