Prereqs:
- [x] https://github.com/ppy/osu-framework/pull/3337

Todo:
- [x] Tests
- [x] Assess ~60B allocation increase.
- [x] Go through the extensions methods with a fine-toothed comb.

There's no breaking changes here :)

`Transform`s can now be given a generic `IEasingFunction`. Example:

```cs
var box = new Box();
box.FadeInFromZero(1000, new MyCustomEasingFunction());

private readonly struct MyCustomEasingFunction : IEasingFunction
{
    public double ApplyEasing(double time)
    {
        return RNG.NextSingle();
    }
}
```

## Benchmarks

Master:

|                  Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| CreateWithDefaultEasing | 2,233.3 ns | 48.24 ns | 45.13 ns | 0.4616 |     - |     - |    2432 B |
|  ApplyWithDefaultEasing |   814.7 ns | 14.88 ns | 13.92 ns | 0.1678 |     - |     - |     880 B |


This PR:

|                    Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
|   CreateWithDefaultEasing | 2,047.3 ns | 13.01 ns | 11.53 ns | 0.4997 |     - |     - |    2624 B |
|    ApplyWithDefaultEasing |   709.4 ns |  2.79 ns |  2.48 ns | 0.1802 |     - |     - |     944 B |
|     CreateWithValueEasing | 2,099.3 ns | 12.14 ns | 10.14 ns | 0.4997 |     - |     - |    2624 B |
|      ApplyWithValueEasing |   707.9 ns |  4.62 ns |  4.10 ns | 0.1802 |     - |     - |     944 B |
| CreateWithReferenceEasing | 3,274.2 ns | 13.85 ns | 12.95 ns | 0.5150 |     - |     - |    2696 B |
|  ApplyWithReferenceEasing | 1,160.9 ns |  3.78 ns |  3.35 ns | 0.1850 |     - |     - |     968 B |