@smoogipoo 2020-03-09
It’s ready for review if the 60B increase is not of concern. Haven’t investigated yet but it’s most likely due to the extra level of generic in transform itself. I’ve been careful to avoid any other boxing.