@smoogipoo 2020-03-09
Let me know if you want me to split up the Interpolation class a bit more, there’s an extra level of generic there for the easing types, otherwise dynamic interpolation function getter fails.