@smoogipoo 2020-03-09
Updated benchmarks:

Master:

|                          Method |         Mean |      Error |     StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |-------------:|-----------:|-----------:|-------:|------:|------:|----------:|
|               CreateSingleBlank |     8.016 ns |  0.1225 ns |  0.1086 ns | 0.0184 |     - |     - |      96 B |
| CreateSequenceWithDefaultEasing | 2,020.451 ns | 17.6533 ns | 16.5129 ns | 0.4997 |     - |     - |    2624 B |
|  ApplySequenceWithDefaultEasing |   720.093 ns |  6.8687 ns |  6.0889 ns | 0.1802 |     - |     - |     944 B |

This branch:

|                            Method |        Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------- |------------:|----------:|----------:|-------:|------:|------:|----------:|
|                 CreateSingleBlank |    12.97 ns |  0.087 ns |  0.073 ns | 0.0184 |     - |     - |      96 B |
|   CreateSequenceWithDefaultEasing | 2,069.82 ns |  8.921 ns |  7.908 ns | 0.4997 |     - |     - |    2624 B |
|    ApplySequenceWithDefaultEasing |   718.42 ns |  4.349 ns |  3.855 ns | 0.1802 |     - |     - |     944 B |
|     CreateSequenceWithValueEasing | 2,063.17 ns | 22.895 ns | 21.416 ns | 0.4997 |     - |     - |    2624 B |
|      ApplySequenceWithValueEasing |   734.37 ns |  4.383 ns |  3.885 ns | 0.1802 |     - |     - |     944 B |
| CreateSequenceWithReferenceEasing | 2,928.66 ns | 16.256 ns | 15.206 ns | 0.5150 |     - |     - |    2696 B |
|  ApplySequenceWithReferenceEasing | 1,006.03 ns | 12.491 ns | 11.073 ns | 0.1850 |     - |     - |     968 B |

So I'm ready to call this complete.